bot_service/pyproject.toml
[project]
name = "bot-service"
version = "0.1.0"
description = "Telegram bot service with JWT auth, Celery, RabbitMQ, Redis, OpenRouter"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "aiogram>=3.10.0",
  "celery>=5.4.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "ruff>=0.14.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-mock>=3.12.0",
  "respx>=0.21.0",
  "fakeredis>=2.23.0",
]

[tool.uv]
dev-dependencies = []